move proto stuff into its own file
